# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.OperatorType:
    enum:
    - IS
    - IS_ONE_OF
    - IS_NOT_ONE_OF
    type: string
    x-enum-varnames:
    - OperatorIS
    - OperatorIsOneOf
    - OperatorIsNotOneOf
...
  dto.ConfigRequest:
    properties:
      apiKey:
        type: string
      authHeaderName:
        type: string
      authType:
        type: string
      autoAnalyze:
        type: boolean
      capabilities:
        items:
          type: string
        type: array
      customHeaders:
        additionalProperties:
          type: string
        type: object
      maxTokens:
        type: integer
      maxToolIterations:
        type: integer
      model:
        type: string
      provider:
        type: string
      url:
        type: string
    required:
    - model
    - provider
    type: object
...
  dto.CreateRuleRequest:
    properties:
      active:
        type: boolean
      agentPlatform:
        type: string
      commands:
        items:
          type: string
        minItems: 1
        type: array
      conditions:
        items:
          $ref: '#/definitions/dto.FilterVM'
        minItems: 1
        type: array
      defaultAgent:
        maxLength: 500
        type: string
      description:
        maxLength: 512
        type: string
      excludedAgents:
        items:
          type: string
        type: array
      id:
        type: integer
      name:
        maxLength: 150
        type: string
      shell:
        maxLength: 20
        type: string
    required:
    - active
    - agentPlatform
    - commands
    - conditions
    - name
    type: object
...
  dto.FilterVM:
    properties:
      field:
        type: string
      operator:
        $ref: '#/definitions/domain.OperatorType'
      value: {}
    type: object
...
  dto.ResolveFilterValuesResponse:
    properties:
      agentPlatform:
        items:
          type: string
        type: array
      users:
        items:
          type: string
        type: array
    type: object
...
  dto.RuleResponse:
    properties:
      active:
        type: boolean
      agentPlatform:
        type: string
      commands:
        items:
          type: string
        type: array
      conditions:
        items:
          $ref: '#/definitions/dto.FilterVM'
        type: array
      defaultAgent:
        type: string
      description:
        type: string
      excludedAgents:
        items:
          type: string
        type: array
      lastModifiedDate:
        type: string
      name:
        type: string
      relPath:
        type: string
      shell:
        type: string
      systemOwner:
        type: boolean
    type: object
...
  dto.ToggleRuleRequest:
    properties:
      enabled:
        type: boolean
    type: object
...
  dto.UpdateRuleRequest:
    properties:
      active:
        type: boolean
      agentPlatform:
        type: string
      commands:
        items:
          type: string
        minItems: 1
        type: array
      conditions:
        items:
          $ref: '#/definitions/dto.FilterVM'
        minItems: 1
        type: array
      defaultAgent:
        maxLength: 500
        type: string
      description:
        maxLength: 512
        type: string
      excludedAgents:
        items:
          type: string
        type: array
      id:
        type: integer
      name:
        maxLength: 150
        type: string
      shell:
        maxLength: 20
        type: string
    required:
    - active
    - agentPlatform
    - commands
    - conditions
    - name
    type: object
...
  github_com_utmstack_utmstack_backend_modules_socai_dto.ConfigResponse:
    properties:
      apiKeySet:
        type: boolean
      authHeaderName:
        type: string
      authType:
        type: string
      autoAnalyze:
        type: boolean
      capabilities:
        items:
          type: string
        type: array
      configured:
        type: boolean
      customHeaders:
        additionalProperties:
          type: string
        type: object
      maxTokens:
        type: integer
      maxToolIterations:
        type: integer
      model:
        type: string
      provider:
        type: string
      url:
        type: string
    type: object
...
info:
  contact: {}
  description: UTMStack backend API — manages IAM, modules, flows, incidents, compliance
    and audit data.
  title: UTMStack Backend Service
  version: "1.0"
paths:
...
  /soar/rules:
    get:
      parameters:
      - description: Page number (0-based)
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      - description: Filter by name (substring)
        in: query
        name: name.contains
        type: string
      - description: Filter by active flag
        in: query
        name: active.equals
        type: boolean
      - description: Filter by agent platform
        in: query
        name: agentPlatform.equals
        type: string
      - description: Filter by system ownership
        in: query
        name: systemOwner.equals
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Total-Count:
              description: Total number of records
              type: integer
          schema:
            items:
              $ref: '#/definitions/dto.RuleResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List alert response rules (flows)
      tags:
      - SOAR Rules
    post:
      consumes:
      - application/json
      description: |-
        Writes the flow to the user overlay as a YAML file. Identity is the
        file relPath; there is no numeric id.
      parameters:
      - description: Flow to create
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RuleResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create alert response rule (flow)
      tags:
      - SOAR Rules
  /soar/rules/{relPath}:
    delete:
      parameters:
      - description: Flow relPath
        in: path
        name: relPath
        required: true
        type: string
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete alert response rule (flow) by relPath
      tags:
      - SOAR Rules
    get:
      parameters:
      - description: Flow relPath
        in: path
        name: relPath
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RuleResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get alert response rule (flow) by relPath
      tags:
      - SOAR Rules
    put:
      consumes:
      - application/json
      parameters:
      - description: Flow relPath
        in: path
        name: relPath
        required: true
        type: string
      - description: Flow to update
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RuleResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update alert response rule (flow)
      tags:
      - SOAR Rules
  /soar/rules/{relPath}/enabled:
    put:
      consumes:
      - application/json
      description: Toggles the .disabled file suffix. Allowed on system flows too.
      parameters:
      - description: Flow relPath
        in: path
        name: relPath
        required: true
        type: string
      - description: Enabled state
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.ToggleRuleRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Updated
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Enable/disable an alert response rule (flow)
      tags:
      - SOAR Rules
  /soar/rules/resolve-filter-values:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResolveFilterValuesResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Resolve filter values for alert response rules
      tags:
      - SOAR Rules
...
  /soc-ai/config:
    get:
      description: Returns the current soc-ai plugin configuration with secrets masked.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_utmstack_utmstack_backend_modules_socai_dto.ConfigResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get SOC AI configuration
      tags:
      - SOC AI
    put:
      consumes:
      - application/json
      description: |-
        Persists the soc-ai plugin configuration (writing the encrypted YAML
        the plugin consumes). Secrets sent as "*****" keep their stored value.
      parameters:
      - description: SOC AI configuration
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/dto.ConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_utmstack_utmstack_backend_modules_socai_dto.ConfigResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update SOC AI configuration
      tags:
      - SOC AI
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"